Add fallback SIGHUP method to ddns
The systemd unit for nsd now explicitly passes `-P ""` which disables
the PID file, even if the configuration file includes it. This change
was made upstream as a result of the need to remove chown'ing the pid
file to nsd for security reasons, and the desire to avoid littering a
pid file on the system.

As systemd is slowly eating the world, everything must be done in
systemd compliant ways, and the old ways are dying out.

I resent this, so here's an ugly hack to keep things working for people
who can't be arsed to fix the systemd unit file to not be dumbshit.

Link to issue discussing this:
https://github.com/NLnetLabs/nsd/issues/317